fix: do not keep types in package entry point
